+++ body/src/main/java/fr/iagl/gamification/constants/ServiceConstants.java
@@ -1,12 +1,19 @@
 package fr.iagl.gamification.constants;
 
+/**
+ * Constante utilisées par les services
+ *
+ * @author Hélène MEYER
+ *
+ */
 public class ServiceConstants {
+	/**
+	 * Constante pour récupérer les utilisateurs ayant un rôle TEACHER
+	 */
 	public static final String CODE_TEACHER = "TEACHER";
 	
 	/**
 	 * La classe ne peut pas être instanciée
 	 */
-	private ServiceConstants(){
-		
-	}
+	private ServiceConstants(){}
 }
