+++ body/src/main/java/fr/iagl/gamification/services/TeacherService.java
@@ -5,9 +5,27 @@ import java.util.List;
 import fr.iagl.gamification.exceptions.GamificationServiceException;
 import fr.iagl.gamification.model.TeacherModel;
 
+/**
+ * Interface du service des professeur
+ *
+ * @author Hélène MEYER
+ *
+ */
 public interface TeacherService {
 
+	/**
+	 * Créer un professeur
+	 * 
+	 * @param teacher le modèle du professeur à ajouter
+	 * @return le professeur créé
+	 * @throws GamificationServiceException si le mail du professeur existe déjà en base de données
+	 */
 	public TeacherModel createTeacher(TeacherModel teacher) throws GamificationServiceException;
 
+	/**
+	 * Récupère tous les professeurs
+	 * 
+	 * @return la liste de tous les professeurs
+	 */
 	public List<TeacherModel> getAllTeacher();
 }
