+++ data/src/main/java/fr/iagl/gamification/repository/UserRepository.java
@@ -13,7 +13,20 @@ import fr.iagl.gamification.entity.UserEntity;
  */
 public interface UserRepository extends CrudRepository<UserEntity, Long>{
 
+	/**
+	 * Donne tous les utilisateurs ayant le rôle de professeur
+	 * 
+	 * @param role le code role
+	 * @return la liste des utilisateurs ayant un code role 'TEACHER'
+	 */
 	public List<UserEntity> findByRole_Role(String role);
 	
-	public UserEntity findByEmail(String email);
+	/**
+	 * Renvoie l'utilisateur 
+	 * 
+	 * @param email email du professeur
+	 * @param role role du professeur
+	 * @return l'utilisateur ayant le rôle professeur et le mail 
+	 */
+	public UserEntity findByEmailAndRole_Role(String email, String role);
 }
